Update deps (webpack@3.0.0).
* About release https://medium.com/webpack/webpack-3-official-release-15fd2dd8f07b

npm WARN babel-loader@7.0.0 requires a peer of webpack@2 but none was installed.
npm WARN extract-text-webpack-plugin@2.1.2 requires a peer of webpack@^2.2.0 but none was installed.
npm WARN html-webpack-plugin@2.28.0 requires a peer of webpack@1 || ^2 || ^2.1.0-beta || ^2.2.0-rc but none was installed.
npm WARN webpack-dev-middleware@1.10.2 requires a peer of webpack@1 || ^2.1.0-beta || ^2.2.0-rc.0 but none was installed.
npm WARN webpack-dev-server@2.4.5 requires a peer of webpack@^2.2.0 but none was installed.